Two finger movement of map enforced on mobile only
If on desktop, can still click and drag, but when on mobile now you must use two fingers to move the map around. One finger scrolling will only scroll the website up and down. This did not require an external library like I expected, but instead uses a built in feature of leaflet.